+++ app/Http/Controllers/PositionsController.php
@@ -5,11 +5,10 @@ namespace App\Http\Controllers;
 use App\Http\DTO\PositionDTO;
 use App\Http\Requests\PaginationRequest;
 use App\Http\Requests\Position\PositionCreateRequest;
+use App\Http\Requests\Position\PositionUpdateRequest;
+use App\Http\Resources\PositionResource;
 use App\Http\Service\PositionService;
 use App\Models\Position;
-use Illuminate\Http\Request;
-use App\Http\Requests\Position\PositionUpdateRequest;
-use Illuminate\Database\Eloquent\ModelNotFoundException;
 
 class PositionsController extends Controller
 {
@@ -20,8 +19,9 @@ class PositionsController extends Controller
 
     public function index(PaginationRequest $request)
     {
+        $this->authorize('viewAny', Position::class);
         $paginator = $request->validated();
-        $response = Position::with('comfortCategories')
+        $positions = Position::with('comfortCategories')
             ->paginate(
                 $paginator['per_page'],
                 ['*'],
@@ -29,33 +29,35 @@ class PositionsController extends Controller
                 $paginator['page']
             );
 
-        return response()->json($response);
+        return PositionResource::collection($positions);
     }
 
     public function store(PositionCreateRequest $request)
     {
+        $this->authorize('create', Position::class);
         $dto = new PositionDTO(
             name: $request->validated('name'),
             comfortCategoryIds: $request->validated('comfort_category'),
         );
-        $createPosition = $this->positionService->createPosition($dto);
-
-        return response()->json($createPosition, 201);
+        $position = $this->positionService->createPosition($dto);
+        return (new PositionResource($position))->response()->setStatusCode(201);
     }
 
     public function update(PositionUpdateRequest $request, Position $position)
     {
+        $this->authorize('update', $position);
         $validated = $request->validated();
         $dto = new PositionDTO(
             name: $validated['name'] ?? $position->name,
             comfortCategoryIds: $validated['comfort_category'] ?? $position->comfortCategories->pluck('id')->toArray(),
         );
-        $update = $this->positionService->updatePosition($position, $dto);
-        return response()->json($update, 200);
+        $position = $this->positionService->updatePosition($position, $dto);
+        return new PositionResource($position);
     }
 
     public function destroy(Position $position)
     {
+        $this->authorize('delete', $position);
         $position->delete();
         return response()->json(null, 200);
     }
